Move testing extension to seperate module
